requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
flask==3.1.0
jinja2==3.1.4
sqlalchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
cryptography==43.0.3
python-dotenv==1.0.1
playwright==1.49.1
httpx==0.28.1
requests==2.32.3
pydantic==2.10.4
python-multipart==0.0.20
